+++ cli/src/cmd_todo.rs
@@ -10,7 +10,8 @@ use crate::{
     todo_formatter::TodoFormatter,
 };
 use aimcal_core::{
-    Aim, Priority, SortOrder, TodoConditions, TodoDraft, TodoPatch, TodoSort, TodoStatus,
+    Aim, LooseDateTime, Priority, SortOrder, TodoConditions, TodoDraft, TodoPatch, TodoSort,
+    TodoStatus,
 };
 use chrono::{Duration, Local};
 use clap::{Arg, ArgMatches, Command, arg};
@@ -56,12 +57,11 @@ impl CmdTodoNew {
         map: &ShortIdMap,
     ) -> Result<(), Box<dyn Error>> {
         log::debug!("Adding new todo...");
-        let due = self
-            .due
-            .as_ref()
-            .map(|a| parse_datetime(a))
-            .transpose()?
-            .flatten();
+        let due = match (self.due, config.default_due) {
+            (Some(due), _) => parse_datetime(&due)?,
+            (None, Some(duration)) => Some(LooseDateTime::Local(Local::now() + duration)),
+            (None, None) => None,
+        };
 
         let draft = TodoDraft {
             uid: Uuid::new_v4().to_string(), // TODO: better uid